requirements.txt
discord.py
dnspython
PyNaCl
async-timeout
python-dotenv
youtube_dl
dblpy
disputils
